refactor(Featured*) change placeholder to fit new sizing
